Fix a bug of getting collected samples in local buffers in ApeX

+++ tf2rl/algos/apex.py
@@ -83,7 +83,7 @@ def explorer(global_rb, queue, trained_steps, is_training_done,
     if n_env > 1:
         kwargs["env_dict"]["priorities"] = {}
     local_rb = ReplayBuffer(**kwargs)
-    local_idx = np.arange(buffer_size)
+    local_idx = np.arange(buffer_size).astype(np.int)
 
     if n_env == 1:
         s = env.reset()
